+++ Osnova.Net/Size.cs
@@ -2,14 +2,26 @@
 
 namespace Osnova.Net
 {
+    /// <summary>
+    /// Sizes of media content
+    /// </summary>
     public class Size
     {
+        /// <summary>
+        /// Width
+        /// </summary>
         [JsonPropertyName("width")]
-        public long Width { get; set; }
+        public int Width { get; set; }
 
+        /// <summary>
+        /// Height
+        /// </summary>
         [JsonPropertyName("height")]
-        public long Height { get; set; }
+        public int Height { get; set; }
 
+        /// <summary>
+        /// Ratio
+        /// </summary>
         [JsonPropertyName("ratio")]
         public double Ratio { get; set; }
     }
